refactor(dictation): one export per file across the module
Five files carried more than one export. Splitting them matches
modules/ai/types, where every file has exactly one.

DictationEngine.ts held five types and WebSpeechRecognition.ts seven;
each type is now its own file. The three factories that declared their
own return type — createDictationEventEmitter, createLivenessWatchdog,
resolveDictationAvailability — import it from types/ instead, which also
puts DictationSurface next to the rest rather than inside the resolver
that happens to consume it.

Behaviour is unchanged: types only, plus the imports that follow them.
