+++ packages/twenty-front/src/modules/ai/dictation/types/DictationEngineListener.ts
@@ -1,0 +1,3 @@
+import { type DictationEngineEvent } from '@/ai/dictation/types/DictationEngineEvent';
+
+export type DictationEngineListener = (event: DictationEngineEvent) => void;
