+++ packages/twenty-front/src/modules/ai/dictation/types/DictationEngineState.ts
@@ -1,0 +1,1 @@
+export type DictationEngineState = 'idle' | 'recording';
