+++ packages/twenty-front/src/modules/ai/dictation/types/DictationEventEmitter.ts
@@ -1,0 +1,8 @@
+import { type DictationEngineEvent } from '@/ai/dictation/types/DictationEngineEvent';
+import { type DictationEngineListener } from '@/ai/dictation/types/DictationEngineListener';
+
+export type DictationEventEmitter = {
+  subscribe: (listener: DictationEngineListener) => () => void;
+  emit: (event: DictationEngineEvent) => void;
+  clear: () => void;
+};
