+++ packages/twenty-front/src/modules/ai/dictation/types/DictationSurface.ts
@@ -1,0 +1,8 @@
+export type DictationSurface = {
+  isIOS: boolean;
+  isStandaloneDisplayMode: boolean;
+  isThirdPartyIOSBrowser: boolean;
+  hasSpeechRecognition: boolean;
+  hasMediaDevices: boolean;
+  isSecureContext: boolean;
+};
