+++ packages/twenty-front/src/modules/ai/dictation/types/LivenessWatchdog.ts
@@ -1,0 +1,5 @@
+export type LivenessWatchdog = {
+  arm: () => void;
+  noteActivity: () => void;
+  disarm: () => void;
+};
