+++ packages/twenty-front/src/modules/ai/dictation/types/WebSpeechRecognitionAlternative.ts
@@ -1,0 +1,3 @@
+export type WebSpeechRecognitionAlternative = {
+  transcript: string;
+};
