+++ packages/twenty-front/src/modules/ai/dictation/types/WebSpeechRecognitionConstructor.ts
@@ -1,0 +1,4 @@
+import { type WebSpeechRecognitionInstance } from '@/ai/dictation/types/WebSpeechRecognitionInstance';
+
+export type WebSpeechRecognitionConstructor =
+  new () => WebSpeechRecognitionInstance;
