+++ packages/twenty-front/src/modules/ai/dictation/types/WebSpeechRecognitionErrorEvent.ts
@@ -1,0 +1,3 @@
+export type WebSpeechRecognitionErrorEvent = {
+  error: string;
+};
