+++ packages/twenty-front/src/modules/ai/dictation/types/WebSpeechRecognitionEvent.ts
@@ -1,0 +1,6 @@
+import { type WebSpeechRecognitionResultList } from '@/ai/dictation/types/WebSpeechRecognitionResultList';
+
+export type WebSpeechRecognitionEvent = {
+  resultIndex: number;
+  results: WebSpeechRecognitionResultList;
+};
