+++ packages/twenty-front/src/modules/ai/dictation/types/WebSpeechRecognitionResult.ts
@@ -1,0 +1,6 @@
+import { type WebSpeechRecognitionAlternative } from '@/ai/dictation/types/WebSpeechRecognitionAlternative';
+
+export type WebSpeechRecognitionResult = {
+  isFinal: boolean;
+  [index: number]: WebSpeechRecognitionAlternative;
+};
