+++ packages/twenty-front/src/modules/ai/dictation/types/WebSpeechRecognitionResultList.ts
@@ -1,0 +1,6 @@
+import { type WebSpeechRecognitionResult } from '@/ai/dictation/types/WebSpeechRecognitionResult';
+
+export type WebSpeechRecognitionResultList = {
+  length: number;
+  [index: number]: WebSpeechRecognitionResult;
+};
